requirements.txt
requests
python-decouple
sqlalchemy
logging
